`sstIterator` can optionally verify the `roachpb.Value` checksum of
encountered values. This patch splits the verification logic out to a
separate `VerifyingMVCCIterator` iterator, which can wrap an arbitrary
`SimpleMVCCIterator`.

This is done because `sstIterator` will shortly be replaced by the
general-purpose `pebbleIterator`, which doesn't support value
verification, and it seems cleaner to implement this logic separately.

Release note: None