@erikgrinaker 2022-06-12
This might not be viable, as the additional overhead turns out to be significant even after optimization attempts:

```
name                                   old time/op  new time/op  delta
SSTIterator/keys=10000/verify=true-24  1.06ms ± 0%  1.22ms ± 4%  +14.67%  (p=0.002 n=8+5)
```

Let's have another look at this once we've also ported the SST iterators over to the new external SST iterators in Pebble (which support range keys and multi-SST iteration), since they come with their own performance overhead.

This may not be that big of an issue in the bigger picture. The only 22.1 hot path that uses verification is `AddSSTable`, but we often don't read/verify the entire SST anyway (only when calculating SST stats).

@dt What's your take on the importance of verification and performance here, given that we often bypass it?