@dt 2022-06-13
This might be more of a question for kv and storage, as to why we even have these per-kv roachpb.Value checksums at all; We already have per-block checksums (using a higher quality checksum at that) which should be telling us that we do are indeed reading back what we wrote, so I don't really know what these are doing for us?