@erikgrinaker 2022-06-13
Looks like we verify them on every remote RPC:

https://github.com/cockroachdb/cockroach/blob/0dade4009b34acb86f29fea22683c71bc68e39c7/pkg/kv/kvclient/kvcoord/transport.go#L211-L218

I suppose that makes sense. So we should really be verifying these values before we put them into SSTs (if they're not from a trusted source), rather than during SST ingestion which, as you say, is already covered by SST checksums.