@erikgrinaker 2022-06-15
I included this in #82799, which already is a pretty severe performance regression, so this additional verification overhead is relatively small. We can discuss either removing verification or optimizing it later, for now we just need something that's functional.